+++ csv_for_d5a/pse2csv_for_d5a_print.h
@@ -14,14 +14,14 @@ void print_headers(
     FILE *f,
     const char *filename,
     int64_t file_offset,
-    uint32_t us_offset,
+    double us_offset,
     pse_record *pr, pse_frame *pf);
 
 void print_pse_spz(
     FILE *f,
     const char *filename,
     int64_t file_offset,
-    uint32_t us_offset,
+    double us_offset,
     pse_record *pr,
     pse_frame *pf,
     int32_t index);
@@ -30,7 +30,7 @@ void print_pse_lpxyz(
     FILE *f,
     const char *filename,
     int64_t file_offset,
-    uint32_t us_offset,
+    double us_offset,
     pse_record *pr,
     pse_frame *pf,
     int32_t index);
@@ -39,7 +39,7 @@ void print_pse_tdxy(
     FILE *f,
     const char *filename,
     int64_t file_offset,
-    uint32_t us_offset,
+    double us_offset,
     pse_record *pr,
     pse_frame *pf);
 
@@ -47,7 +47,7 @@ void print_pse_tdzi(
     FILE *f,
     const char *filename,
     int64_t file_offset,
-    uint32_t us_offset,
+    double us_offset,
     pse_record *pr,
     pse_frame *pf);
 
@@ -55,7 +55,7 @@ void print_pse_meta(
     FILE *f,
     const char *filename,
     int64_t file_offset,
-    uint32_t us_offset,
+    double us_offset,
     int32_t record_no,
     int32_t frame_no,
     pse_record *pr,
