+++ csv_for_d5a/wtn2csv_for_d5a_print.h
@@ -16,14 +16,14 @@ void print_headers(
     FILE *f,
     const char *filename,
     int64_t file_offset,
-    uint32_t us_offset,
+    double us_offset,
     wtn_record *wnr, wtn_frame *wnf);
 
 void print_wtn_spz(
     FILE *f,
     const char *filename,
     int64_t file_offset,
-    uint32_t us_offset,
+    double us_offset,
     wtn_record *wnr,
     wtn_frame *wnf,
     int32_t index);
@@ -32,7 +32,7 @@ void print_wtn_lsg(
     FILE *f,
     const char *filename,
     int64_t file_offset,
-    uint32_t us_offset,
+    double us_offset,
     wtn_record *wnr,
     wtn_frame *wnf,
     int32_t index);
@@ -41,7 +41,7 @@ void print_wtn_lpxyz(
     FILE *f,
     const char *filename,
     int64_t file_offset,
-    uint32_t us_offset,
+    double us_offset,
     wtn_record *wnr,
     wtn_frame *wnf,
     int32_t index);
@@ -50,7 +50,7 @@ void print_wtn_tdxy(
     FILE *f,
     const char *filename,
     int64_t file_offset,
-    uint32_t us_offset,
+    double us_offset,
     wtn_record *wnr,
     wtn_frame *wnf);
 
@@ -58,7 +58,7 @@ void print_wtn_tdzi(
     FILE *f,
     const char *filename,
     int64_t file_offset,
-    uint32_t us_offset,
+    double us_offset,
     wtn_record *wnr,
     wtn_frame *wnf);
 
@@ -66,7 +66,7 @@ void print_wtn_meta(
     FILE *f,
     const char *filename,
     int64_t file_offset,
-    uint32_t us_offset,
+    double us_offset,
     int32_t record_no,
     int32_t frame_no,
     wtn_record *wnr,
